fix хапуска приложения
