+++ package.json
@@ -4,9 +4,9 @@
   "main": "main.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-  "serve": "http-server -p 3000 -a localhost ./",
-  "start": "concurrently \"npm run serve\" \"npm run electron\"",
-  "electron": "wait-on http://localhost:3000 && electron ."
+    "serve": "http-server -p 3000 -a localhost ./",
+    "start": "concurrently \"npm run serve\" \"npm run electron\"",
+    "electron": "wait-on http://localhost:3000 && electron --enabled-logging ."
   },
   "keywords": [],
   "author": "",
